get frontend ready for the 1st GA
