+++ backend/src/resource/asset/index.html
@@ -8,6 +8,6 @@
         <link rel="stylesheet" href="/asset/toastui-editor.min-f34c71ba161ffc6f.css">
         
         
-    <script type="module">import init from '/asset/index-99090a8124021b85.js';init('/asset/index-99090a8124021b85_bg.wasm');</script></head>
+    <script type="module">import init from '/asset/index-c2869c2a5baead37.js';init('/asset/index-c2869c2a5baead37_bg.wasm');</script></head>
     <body>
 </body></html>
