+++ backend/src/service/asset_list.rs
@@ -3,27 +3,12 @@
 ("bootstrap.min-6c05f7396999b9df.css", include_bytes!(r#"..\resource\asset/bootstrap.min-6c05f7396999b9df.css.gz"#), "text/css"),
 ("bootstrap.min.css.map", include_bytes!(r#"..\resource\asset/bootstrap.min.css.map.gz"#), "text/css"),
 ("codemirror.min-79bbbb1069b6fc3d.css", include_bytes!(r#"..\resource\asset/codemirror.min-79bbbb1069b6fc3d.css.gz"#), "text/css"),
-("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
 ("fonts/bootstrap-icons.woff", include_bytes!(r#"..\resource\asset/fonts\bootstrap-icons.woff.gz"#), ""),
 ("fonts/bootstrap-icons.woff2", include_bytes!(r#"..\resource\asset/fonts\bootstrap-icons.woff2.gz"#), ""),
-("index-30e0846c90aa8f94.js", include_bytes!(r#"..\resource\asset/index-30e0846c90aa8f94.js.gz"#), "text/javascript"),
-("index-30e0846c90aa8f94_bg.wasm", include_bytes!(r#"..\resource\asset/index-30e0846c90aa8f94_bg.wasm.gz"#), "application/wasm"),
-("index-4ddbde520bc9e6e2.js", include_bytes!(r#"..\resource\asset/index-4ddbde520bc9e6e2.js.gz"#), "text/javascript"),
-("index-4ddbde520bc9e6e2_bg.wasm", include_bytes!(r#"..\resource\asset/index-4ddbde520bc9e6e2_bg.wasm.gz"#), "application/wasm"),
-("index-71b5e40f0d8ed9d5.js", include_bytes!(r#"..\resource\asset/index-71b5e40f0d8ed9d5.js.gz"#), "text/javascript"),
-("index-71b5e40f0d8ed9d5_bg.wasm", include_bytes!(r#"..\resource\asset/index-71b5e40f0d8ed9d5_bg.wasm.gz"#), "application/wasm"),
-("index-7d6ddff66ba08a77.js", include_bytes!(r#"..\resource\asset/index-7d6ddff66ba08a77.js.gz"#), "text/javascript"),
-("index-7d6ddff66ba08a77_bg.wasm", include_bytes!(r#"..\resource\asset/index-7d6ddff66ba08a77_bg.wasm.gz"#), "application/wasm"),
-("index-99090a8124021b85.js", include_bytes!(r#"..\resource\asset/index-99090a8124021b85.js.gz"#), "text/javascript"),
-("index-99090a8124021b85_bg.wasm", include_bytes!(r#"..\resource\asset/index-99090a8124021b85_bg.wasm.gz"#), "application/wasm"),
-("index-dbad60a00b71da94.js", include_bytes!(r#"..\resource\asset/index-dbad60a00b71da94.js.gz"#), "text/javascript"),
-("index-dbad60a00b71da94_bg.wasm", include_bytes!(r#"..\resource\asset/index-dbad60a00b71da94_bg.wasm.gz"#), "application/wasm"),
-("index-e471e8fae1734f5c.js", include_bytes!(r#"..\resource\asset/index-e471e8fae1734f5c.js.gz"#), "text/javascript"),
-("index-e471e8fae1734f5c_bg.wasm", include_bytes!(r#"..\resource\asset/index-e471e8fae1734f5c_bg.wasm.gz"#), "application/wasm"),
+("index-c2869c2a5baead37.js", include_bytes!(r#"..\resource\asset/index-c2869c2a5baead37.js.gz"#), "text/javascript"),
+("index-c2869c2a5baead37_bg.wasm", include_bytes!(r#"..\resource\asset/index-c2869c2a5baead37_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
 ("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
-("toastui-editor-all.min-4e0cc8f1c6f6c721.js", include_bytes!(r#"..\resource\asset/toastui-editor-all.min-4e0cc8f1c6f6c721.js.gz"#), "text/javascript"),
 ("toastui-editor-all.min.js", include_bytes!(r#"..\resource\asset/toastui-editor-all.min.js.gz"#), "text/javascript"),
 ("toastui-editor.min-f34c71ba161ffc6f.css", include_bytes!(r#"..\resource\asset/toastui-editor.min-f34c71ba161ffc6f.css.gz"#), "text/css"),
-("toastui-editor.min.css", include_bytes!(r#"..\resource\asset/toastui-editor.min.css.gz"#), "text/css"),
 ]
